** Added support for updating workspace type via REST
