+++ muikku-core/src/main/java/fi/otavanopisto/muikku/schooldata/entity/WorkspaceType.java
@@ -1,8 +1,10 @@
 package fi.otavanopisto.muikku.schooldata.entity;
 
+import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
+
 public interface WorkspaceType extends SchoolDataEntity {
 	
-	public String getIdentifier();
+	public SchoolDataIdentifier getIdentifier();
 	
 	public String getName();
